SCRUM-5 refactoring layout
