+++ src/features/quote/model/quotesReducer.ts
@@ -1,7 +1,7 @@
 import { quoteApi } from "../api/quoteApi"
 import { Author, Quote } from "@/features/quote/api/quoteApi.types"
 import { AuthorAndQuote } from "@/features/profile/api/profileApi.types"
-import { resetAllDataAC } from "@/features/auth/model/authSlice"
+import { resetAllDataAC } from "@/features/auth/model/authReducer"
 
 
 const initialState = {
